DynamicTablesPkg/Smbios: Fix X64 build warnings
Moving the standalone SMBIOS generators to the common component section
adds them to the X64 build. VS2022 reports narrowing warnings for native-
width, wide, or enum values converted to fixed-width SMBIOS fields. These
warnings are treated as errors under /WX.

Use UINTN for the Type 7 table-list count and index. Add bounds checks for
the Type 4 processor counts and Type 7 cache level where CM-derived values
could exceed the corresponding SMBIOS fields, and make the remaining
intended conversions explicit in the Type 4, Type 7, Type 16, Type 17,
Type 19, and Type 44 generators.

This rejects out-of-range processor counts and cache levels instead of
silently truncating them, while preserving the existing behavior for valid
configuration data.

+++ DynamicTablesPkg/Library/Smbios/SmbiosType16Lib/SmbiosType16Generator.c
@@ -108,7 +108,7 @@ UpdateSmbiosType16Size (
   SizeKb = SizeBytes / SIZE_1KB;
 
   if (SizeBytes < EXTENDED_SIZE_THRESHOLD) {
-    SmbiosRecord->MaximumCapacity = SizeKb;
+    SmbiosRecord->MaximumCapacity = (UINT32)SizeKb;
   } else {
     SmbiosRecord->MaximumCapacity         = 0x80000000;
     SmbiosRecord->ExtendedMaximumCapacity = SizeBytes;
